changes made

+++ Android/PlantProtector/app/src/main/java/com/example/plantprotector/SecondFragment.java
@@ -49,8 +49,8 @@ public class SecondFragment extends Fragment {
         if (((MainActivity)getActivity()).isRunning()) {
             if (plantInfo.message != null){
                 ((TextView)getView().findViewById(R.id.plant_1_status)).setText(plantInfo.message);
-                ((TextView)getView().findViewById(R.id.plant_1_latency)).setText((int) plantInfo.latency);
-                ((TextView)getView().findViewById(R.id.plant_1_packet_loss)).setText((int) plantInfo.packetLoss);
+                ((TextView)getView().findViewById(R.id.plant_1_latency)).setText(plantInfo.latency);
+                ((TextView)getView().findViewById(R.id.plant_1_packet_loss)).setText(plantInfo.packetLoss);
             } else {
                 ((TextView)getView().findViewById(R.id.plant_1_status)).setText(R.string.updating);
                 ((TextView)getView().findViewById(R.id.plant_1_latency)).setText(R.string.updating);
